[Feat] 맵 초안 제작
게임에 사용될 맵을 1F~4F 까지로 만들었다. 일단 룸은 당장은 2개가 구현되어있고 각 층별로 던전을 랜덤하게 구성한다. 4F는 보스방으로 구성했다. 각 층의 끝에 있는 계단으로 이동하면 자동으로 다음층으로 이동하도록 했다.

+++ Assets/Student/WSY/Monster.cs
@@ -114,7 +114,7 @@ public class Monster : MonoBehaviour, IDamagable
             animator.SetTrigger("Die");
 
             // 플레이어의 골드 증가시키기
-            Manager.Game.AddGold(gold * Random.Range(0.8f, 1.1f));
+            Manager.Game.AddGold((int)(gold * Random.Range(0.8f, 1.1f)));
 
             // 플레이어의 경험치 증가시키기 (데이터 매니저 통해서)
             Manager.Data.playerStatus.AddExp(exp);
